+++ main/features/org.talend.hadoop.distribution.feature/feature.xml
@@ -51,5 +51,5 @@
   <plugin id="org.talend.hadoop.distribution.piv200" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
   <plugin id="org.talend.hadoop.distribution.qubole" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
   <plugin id="org.talend.hadoop.distribution.spl30x" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
-  <plugin id="org.talend.spark.distribution.spark311" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
+  <plugin id="org.talend.spark.distribution.spark31x" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
 </feature>
