+++ main/plugins/org.talend.spark.distribution.spark31x/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.talend.spark.distribution.spark311</name>
+	<name>org.talend.spark.distribution.spark31x</name>
 	<comment></comment>
 	<projects>
 	</projects>
