+++ main/plugins/org.talend.spark.distribution.spark31x/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Spark311
-Bundle-SymbolicName: org.talend.spark.distribution.spark311;singleton:=true
+Bundle-SymbolicName: org.talend.spark.distribution.spark31x;singleton:=true
 Bundle-Version: 7.4.1.qualifier
 Bundle-Vendor: .Talend SA.
 Service-Component: OSGI-INF/spark311.xml
@@ -9,4 +9,4 @@ Bundle-ActivationPolicy: lazy
 Require-Bundle: org.talend.core.runtime,
  org.talend.hadoop.distribution
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: org.talend.spark.distribution.spark311
+Export-Package: org.talend.spark.distribution.spark31x
