+++ main/plugins/org.talend.spark.distribution.spark31x/OSGI-INF/spark311.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
-<component name="Spark311Distribution">
-  <implementation class="org.talend.spark.distribution.spark311.Spark311Distribution"/>
+<component name="Spark31xDistribution">
+  <implementation class="org.talend.spark.distribution.spark31x.Spark31xDistribution"/>
   <service>
     <provide interface="org.talend.hadoop.distribution.component.HadoopComponent"/>
     <provide interface="org.talend.hadoop.distribution.component.HiveOnSparkComponent"/>
