+++ main/plugins/org.talend.spark.distribution.spark31x/build.properties
